refactor(mcp): consolidate all MCP types to centralized locations
• Extract 25+ interfaces from local MCP files to src/lib/types/mcpTypes.ts (763 lines)
• Move ToolImplementation, ToolExecutionOptions to src/lib/types/tools.ts
• Enhance MCPServerCategory with deployment + application domain categories
• Update imports across 20 files: toolRegistry, factory, circuitBreaker, servers
• Convert mcpContract.ts to re-export file for backward compatibility
• Resolve type signature compatibility issues
• Zero functional changes - pure type organization refactor
