+++ .gitignore
@@ -112,6 +112,9 @@ test-directTools-temp/
 site/
 _site/
 quality-metrics.json
+
+# Claude Code
+.claude/
 .claude/settings.local.json
 
 # Security Files
