+++ src/lib/mcp/registry.ts
@@ -2,31 +2,10 @@
  * MCP Registry - Industry Standard Interface with camelCase
  */
 
-import type {
-  DiscoveredMcp,
-  ExecutionContext,
-  ToolInfo,
-} from "./contracts/mcpContract.js";
+import type { DiscoveredMcp, McpRegistry } from "../types/mcpTypes.js";
+import type { ToolInfo, ExecutionContext } from "../types/tools.js";
 import type { UnknownRecord } from "../types/common.js";
 import { registryLogger } from "../utils/logger.js";
-
-/**
- * MCP Registry interface with optional methods for maximum flexibility
- */
-export interface McpRegistry {
-  // All methods optional (maximum flexibility)
-  registerServer?(
-    serverId: string,
-    serverConfig?: unknown,
-    context?: ExecutionContext,
-  ): Promise<void>;
-  executeTool?<T = unknown>(
-    toolName: string,
-    args?: unknown,
-    context?: ExecutionContext,
-  ): Promise<T>;
-  listTools?(context?: ExecutionContext): Promise<ToolInfo[]>;
-}
 
 /**
  * Simple MCP registry for plugin management
