+++ src/lib/mcp/servers/aiProviders/aiCoreServer.ts
@@ -7,7 +7,10 @@
 import { z } from "zod";
 import type { Unknown } from "../../../types/common.js";
 import { createMCPServer } from "../../factory.js";
-import type { NeuroLinkExecutionContext, ToolResult } from "../../factory.js";
+import type {
+  NeuroLinkExecutionContext,
+  ToolResult,
+} from "../../../types/mcpTypes.js";
 import {
   getBestProvider,
   getAvailableProviders,
