+++ src/lib/mcp/servers/aiProviders/aiWorkflowTools.ts
@@ -9,9 +9,9 @@ import type {
   NeuroLinkMCPTool,
   NeuroLinkExecutionContext,
   ToolResult,
-} from "../../factory.js";
-import { AIProviderFactory } from "../../../core/factory.js";
+} from "../../../types/mcpTypes.js";
 import type { AIProvider } from "../../../types/index.js";
+import { AIProviderFactory } from "../../../core/factory.js";
 import { getBestProvider } from "../../../utils/providerUtils.js";
 import { DEFAULT_MAX_TOKENS } from "../../../core/constants.js";
